+++ src/cloudly/gcp/alloydb.py
@@ -1,0 +1,2 @@
+# AlloyDB supports ScaNN vector indexing. See
+# https://cloud.google.com/blog/products/databases/understanding-the-scann-index-in-alloydb