+++ tests/gcp/test_gcp_sql.py
@@ -1,12 +1,12 @@
 from uuid import uuid4
 
-from cloudly.gcp.cloudsql.postgres import Instance
+from cloudly.gcp.sql import PostgresInstance
 
 REGION = 'us-west1'
 
 
 def test_instance():
-    inst = Instance.create(
+    inst = PostgresInstance.create(
         name=f"test-pg-{str(uuid4()).split('-')[0]}",
         region=REGION,
         root_password='rootuser',  # noqa: S106
@@ -15,11 +15,10 @@ def test_instance():
     print(inst)
     try:
         print('name:', inst.name)
-        assert inst.name in [v.name for v in Instance.list()]
+        assert inst.name in [v.name for v in PostgresInstance.list()]
         print('replicas:', inst.replica_names)
         print('instance_type:', inst.instance_type)
         print('url:', inst.url)
         print('host_ip:', inst.host_ip)
     finally:
-        # inst.delete()
-        pass
+        inst.delete()